% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\docType{data}
\name{FNparameters}
\alias{FNparameters}
\title{The application server-side}
\format{
An object of class \code{character} of length 1.
}
\usage{
FNparameters
}
\arguments{
\item{request}{Internal parameter for `{shiny}`. 
DO NOT REMOVE.}
}
\description{
The application server-side
}
\keyword{datasets}
