% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{plot_prep_ui}
\alias{plot_prep_ui}
\title{The application server-side}
\usage{
plot_prep_ui(id)
}
\arguments{
\item{request}{Internal parameter for `{shiny}`. 
DO NOT REMOVE.}
}
\description{
The application server-side
}
